requirements.txt
requests>=2.31.0
feedparser>=6.0.10
openai>=1.40.0
python-dotenv>=1.0.0
pytz>=2023.3